For future-proofing, changed C:/Users/Elliott to C:/Users/%A_UserName% in relevant files.
This should help avoid major headaches down the line if, say, I'm getting a new PC for a job, etc. Just something I won't have to change/deal with later. As long as the file/folder structure is identical to how it is on my main PC, it should all work the same.