I made it so that MusicBee NumPad mode has no toggles, and normal mode has just NumLock.
There are two reasons why I switched it to no longer having any toggles. The first being it's the most commonly used mode, and the light being constantly on is nasty to look at. Second, if I ever need to use Normal Mode, it's more logical to have NumLock enabled for that, since basically everyone else does that.